+++ tests/example.test.js
@@ -18,16 +18,3 @@ describe('Asserts', () => {
     console.log(manager.contracts['15_MessageReceiver'].address);
   });
 });
-
-export async function deployCheck(address, client) {
-  while (true) {
-    await new Promise((resolve) => setTimeout(resolve, 500));
-    _contract = await client.queries.accounts.query(
-      {
-        id: { eq: address },
-      },
-      'id'
-    );
-    if (_contract.length > 0) break;
-  }
-}